Step 1 of refactor test, saving progress in Dev Branch.
